Finish command bar work
Fixes #67.
Fixes #68.
Fixes #75.

IT'S FINALLY DONE!  All tools now save/restore last-used settings.  All
support saving/loading presets.  All have a "reset to default" option,
and all can be randomized.  This was a huge task, but one well worth it.
Program-wide usability has been significantly enhanced by it.

Two holdouts are the Resize and Canvas Resize tools.  These are a bit
tricky, as we don't want them automatically loading last-used
settings... so I'll need to add a new property to the command bar to
handle that possibility.  I'll have that done in the next few days.

+++ Modules/VBP_ProcessorModule.bas
@@ -1013,7 +1013,7 @@ Public Sub Process(ByVal processID As String, Optional showDialog As Boolean = F
             If showDialog Then
                 FormTwins.Show vbModal, FormMain
             Else
-                FormTwins.GenerateTwins cParams.GetByte(1)
+                FormTwins.GenerateTwins cParams.GetLong(1)
             End If
             
         Case "Vignetting"
